refactor/#88 : mv mainMissions to overview package
