+++ app/src/main/java/com/kiero/presentation/main/navigation/MainAppState.kt
@@ -35,7 +35,7 @@ import com.kiero.presentation.parent.navigation.ParentMission
 import com.kiero.presentation.parent.navigation.Reward
 import com.kiero.presentation.parent.navigation.Schedule
 import com.kiero.presentation.parent.screen.mission.directadd.navigation.navigateToMissionAdd
-import com.kiero.presentation.parent.screen.mission.navigation.navigateToParentMission
+import com.kiero.presentation.parent.screen.mission.overview.navigation.navigateToParentMission
 import com.kiero.presentation.parent.screen.reward.navigation.navigateToReward
 import com.kiero.presentation.parent.screen.schedule.navigation.navigateToSchedule
 import com.kiero.presentation.parent.screen.schedule.plan.navigation.navigateToScheduleAdd
