+++ app/src/main/java/com/kiero/presentation/parent/navigation/ParentNavigation.kt
@@ -9,7 +9,7 @@ import androidx.navigation.compose.navigation
 import com.kiero.core.navigation.Route
 import com.kiero.presentation.parent.screen.alarm.navigation.parentAlarmNavGraph
 import com.kiero.presentation.parent.screen.journey.navigation.parentJourneyNavGraph
-import com.kiero.presentation.parent.screen.mission.navigation.parentMissionNavGraph
+import com.kiero.presentation.parent.screen.mission.overview.navigation.parentMissionNavGraph
 import com.kiero.presentation.parent.screen.mission.autoadd.navigation.parentAutoMissionAddNavGraph
 import com.kiero.presentation.parent.screen.mission.directadd.navigation.parentMissionAddNavGraph
 import com.kiero.presentation.parent.screen.mypage.navigation.parentMypageNavGraph
