--- app/src/main/java/com/kiero/presentation/parent/screen/mission/component/datepicker/component/DaysOfWeekTitle.kt
@@ -1,67 +1,0 @@
-package com.kiero.presentation.parent.screen.mission.component.datepicker.component
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.aspectRatio
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.key
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.kiero.core.designsystem.theme.KieroTheme
-import com.kiero.presentation.parent.screen.mission.component.datepicker.util.daysOfWeek
-import kotlinx.collections.immutable.ImmutableList
-import kotlinx.collections.immutable.toImmutableList
-import java.time.DayOfWeek
-import java.time.format.TextStyle
-import java.util.Locale
-
-@Composable
-fun DaysOfWeekTitle(
-    daysOfWeek: ImmutableList<DayOfWeek>,
-    modifier: Modifier = Modifier,
-) {
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .background(
-                color = KieroTheme.colors.gray900,
-            ),
-        horizontalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        daysOfWeek.forEach { dayOfWeek ->
-
-            key(dayOfWeek) {
-                Box(
-                    modifier = Modifier
-                        .weight(1f)
-                        .aspectRatio(1f),
-                    contentAlignment = Alignment.Center
-                ) {
-                    Text(
-                        text = dayOfWeek.getDisplayName(
-                            TextStyle.SHORT,
-                            Locale.getDefault()
-                        ),
-                        color = KieroTheme.colors.white,
-                        style = KieroTheme.typography.regular.body3
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun DaysOfWeekTitlePreview() {
-    KieroTheme {
-        val daysOfWeek = daysOfWeek(firstDayOfWeek = DayOfWeek.SUNDAY).toImmutableList()
-        DaysOfWeekTitle(daysOfWeek = daysOfWeek)
-    }
-}
