--- app/src/main/java/com/kiero/presentation/parent/screen/mission/component/datepicker/model/CalendarMonth.kt
@@ -1,11 +1,0 @@
-package com.kiero.presentation.parent.screen.mission.component.datepicker.model
-
-import androidx.compose.runtime.Immutable
-import kotlinx.collections.immutable.ImmutableList
-import java.time.YearMonth
-
-@Immutable
-data class CalendarMonth(
-    val yearMonth: YearMonth,
-    val weekDays: ImmutableList<ImmutableList<CalendarDay>>
-)
