--- app/src/main/java/com/kiero/presentation/parent/screen/mission/component/datepicker/model/DownTimeStatus.kt
@@ -1,8 +1,0 @@
-package com.kiero.presentation.parent.screen.mission.component.datepicker.model
-
-
-enum class DateState {
-    PAST,
-    TODAY,
-    FUTURE
-}
