--- app/src/main/java/com/kiero/presentation/parent/screen/mission/component/datepicker/util/CalendarCalCulator.kt
@@ -1,5 +1,0 @@
-package com.kiero.presentation.parent.screen.mission.component.datepicker.util
-
-import java.time.DayOfWeek
-
-fun DayOfWeek.daysUntil(other: DayOfWeek) = (other.value - value + 7) % 7
