+++ app/src/main/java/com/kiero/presentation/parent/screen/mission/overview/component/datepicker/model/DayStatus.kt
@@ -1,0 +1,12 @@
+package com.kiero.presentation.parent.screen.mission.overview.component.datepicker.model
+
+import java.time.LocalDate
+
+fun getDateState(date: LocalDate, today: LocalDate = LocalDate.now()): DateState {
+    return when {
+        date.isBefore(today) -> DateState.PAST
+        date.isEqual(today) -> DateState.TODAY
+        else -> DateState.FUTURE
+    }
+}
+
