+++ app/src/main/java/com/kiero/presentation/parent/screen/mission/overview/component/datepicker/model/DownTimeStatus.kt
@@ -1,0 +1,8 @@
+package com.kiero.presentation.parent.screen.mission.overview.component.datepicker.model
+
+
+enum class DateState {
+    PAST,
+    TODAY,
+    FUTURE
+}
