+++ app/src/main/java/com/kiero/presentation/parent/screen/mission/overview/component/datepicker/util/CalendarCalCulator.kt
@@ -1,0 +1,5 @@
+package com.kiero.presentation.parent.screen.mission.overview.component.datepicker.util
+
+import java.time.DayOfWeek
+
+fun DayOfWeek.daysUntil(other: DayOfWeek) = (other.value - value + 7) % 7
